revert before javafx; gradle 8.5; java 11; reformat; io.mattw.jpowder
