+++ build.gradle
@@ -1,16 +1,73 @@
-apply plugin: 'java'
-apply plugin: 'application'
-apply plugin: 'eclipse'
-apply plugin: 'idea'
-sourceCompatibility = 1.5
-version = '0.0.1'
-mainClassName = "main.java.powder.Window"
-repositories {
-    mavenCentral()
-}
-jar {
-    manifest {
-        attributes "Main-Class": "$mainClassName"
+buildscript {
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+    }
+    dependencies {
+        classpath "gradle.plugin.com.github.johnrengelman:shadow:7.1.2"
     }
 }
 
+plugins {
+    id 'com.github.johnrengelman.shadow' version '7.1.2'
+    id "application"
+}
+
+repositories {
+    mavenCentral()
+}
+
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(11)
+        vendor = JvmVendorSpec.ADOPTOPENJDK
+    }
+}
+
+jar {
+    manifest {
+        // fix for shadow plugin + log4j
+        attributes 'Multi-Release': 'true'
+    }
+}
+
+def buildTime() {
+    return new Date().format('yyyyMMdd.HHmmss')
+}
+
+task packageJar(type: Copy) {
+    delete layout.buildDirectory.dir('package')
+    dependsOn('shadowJar')
+
+    from layout.projectDirectory.file('README.md')
+    into layout.buildDirectory.dir('package')
+
+    from layout.projectDirectory.file('LICENSE')
+    into layout.buildDirectory.dir('package')
+
+    from layout.buildDirectory.file('libs/jpowder-game-all.jar')
+    into layout.buildDirectory.dir('package')
+
+    def version = buildTime();
+    from layout.buildDirectory.file('package/jpowder-game-all.jar')
+    rename { String fileName -> fileName.replace('all', version) }
+}
+
+// Setup application, dependencies
+group = "io.mattw.jpowder"
+
+application {
+    mainClassName = "io.mattw.jpowder.Window"
+}
+
+dependencies {
+    implementation 'commons-io:commons-io:2.15.1'
+    implementation 'org.apache.logging.log4j:log4j-api:2.23.1'
+    implementation 'org.apache.logging.log4j:log4j-core:2.23.1'
+
+    // UI
+    implementation 'com.formdev:flatlaf:3.4'
+    implementation 'com.formdev:flatlaf-extras:3.4'
+    implementation 'com.formdev:flatlaf-intellij-themes:3.4'
+    implementation 'com.miglayout:miglayout-swing:11.3'
+}
