+++ src/main/java/io/mattw/jpowder/Item.java
@@ -1,0 +1,10 @@
+package io.mattw.jpowder;
+
+import java.awt.*;
+
+public class Item {
+    public int id = 0;
+    public String name = "Item";
+    public String description = "Description";
+    public Color color = Color.GRAY;
+}
