+++ src/main/java/io/mattw/jpowder/elements/ElementMovement.java
@@ -1,0 +1,7 @@
+package io.mattw.jpowder.elements;
+
+import io.mattw.jpowder.particles.Particle;
+
+public interface ElementMovement {
+    void move(Particle p);
+}
