--- src/main/java/powder/Item.java
@@ -1,10 +1,0 @@
-package main.java.powder;
-
-import java.awt.Color;
-
-public class Item {
-	public int id = 0;
-	public String name = "Item";
-	public String description = "Description";
-	public Color color = Color.GRAY;
-}
