--- src/main/java/powder/elements/ElementMovement.java
@@ -1,7 +1,0 @@
-package main.java.powder.elements;
-
-import main.java.powder.particles.Particle;
-
-public interface ElementMovement {
-    void move(Particle p);
-}
